created make_query_document which is simpler than attempting to backtrack
